13.2: Use Int64 for coordinates & derrived values (such as cost)

+++ day_13/src/combination.cr
@@ -1,6 +1,6 @@
 record Combination,
   current_point : Point,
-  total_cost : Int32 do
+  total_cost : Int64 do
 
   def +(action : Action)
     copy_with(
